myscripts/web/backend/requirements.txt
fastapi
uvicorn
pandas
xarray
google-cloud-storage 